+++ rpc/filter_api.go
@@ -1,12 +1,9 @@
 package rpc
 
 import (
-	"encoding/json"
-	"fmt"
-	"math/big"
+	"errors"
 
 	"github.com/cosmos/cosmos-sdk/client/context"
-	"github.com/cosmos/ethermint/x/evm/types"
 
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/eth/filters"
@@ -62,11 +59,14 @@ func (e *PublicFilterAPI) UninstallFilter(id rpc.ID) bool {
 // If the filter is a block filter, it returns an array of block hashes.
 // If the filter is a pending transaction filter, it returns an array of transaction hashes.
 func (e *PublicFilterAPI) GetFilterChanges(id rpc.ID) (interface{}, error) {
+	if e.filters[id] == nil {
+		return nil, errors.New("invalid filter ID")
+	}
 	return e.filters[id].getFilterChanges()
 }
 
 // GetFilterLogs returns an array of all logs matching filter with given id.
-func (e *PublicFilterAPI) GetFilterLogs(id rpc.ID) []*ethtypes.Log {
+func (e *PublicFilterAPI) GetFilterLogs(id rpc.ID) ([]*ethtypes.Log, error) {
 	return e.filters[id].getFilterLogs()
 }
 
@@ -74,50 +74,6 @@ func (e *PublicFilterAPI) GetFilterLogs(id rpc.ID) []*ethtypes.Log {
 //
 // https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_getlogs
 func (e *PublicFilterAPI) GetLogs(criteria filters.FilterCriteria) ([]*ethtypes.Log, error) {
-	var filter *Filter
-	if criteria.BlockHash != nil {
-		/*
-			Still need to add blockhash in prepare function for log entry
-		*/
-		filter = NewFilterWithBlockHash(e.backend, &criteria)
-		results := e.getLogs()
-		logs := filterLogs(results, nil, nil, filter.addresses, filter.topics)
-		return logs, nil
-	}
-	// Convert the RPC block numbers into internal representations
-	begin := rpc.LatestBlockNumber.Int64()
-	if criteria.FromBlock != nil {
-		begin = criteria.FromBlock.Int64()
-	}
-	from := big.NewInt(begin)
-	end := rpc.LatestBlockNumber.Int64()
-	if criteria.ToBlock != nil {
-		end = criteria.ToBlock.Int64()
-	}
-	to := big.NewInt(end)
-	results := e.getLogs()
-	logs := filterLogs(results, from, to, criteria.Addresses, criteria.Topics)
-
-	return returnLogs(logs), nil
-}
-
-func (e *PublicFilterAPI) getLogs() (results []*ethtypes.Log) {
-	l, _, err := e.cliCtx.QueryWithData(fmt.Sprintf("custom/%s/logs", types.ModuleName), nil)
-	if err != nil {
-		fmt.Printf("error from querier %e ", err)
-	}
-
-	if err := json.Unmarshal(l, &results); err != nil {
-		panic(err)
-	}
-	return results
-}
-
-// returnLogs is a helper that will return an empty log array in case the given logs array is nil,
-// otherwise the given logs array is returned.
-func returnLogs(logs []*ethtypes.Log) []*ethtypes.Log {
-	if logs == nil {
-		return []*ethtypes.Log{}
-	}
-	return logs
+	filter := NewFilter(e.backend, &criteria)
+	return filter.getFilterLogs()
 }