+++ rpc/filters.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/eth/filters"
+	"github.com/ethereum/go-ethereum/log"
 )
 
 /*
@@ -37,7 +38,7 @@ type Filter struct {
 
 // NewFilter returns a new Filter
 func NewFilter(backend Backend, criteria *filters.FilterCriteria) *Filter {
-	return &Filter{
+	filter := &Filter{
 		backend:   backend,
 		fromBlock: criteria.FromBlock,
 		toBlock:   criteria.ToBlock,
@@ -46,6 +47,8 @@ func NewFilter(backend Backend, criteria *filters.FilterCriteria) *Filter {
 		typ:       logFilter,
 		stopped:   false,
 	}
+
+	return filter
 }
 
 // NewFilterWithBlockHash returns a new Filter with a blockHash.
@@ -139,28 +142,107 @@ func (f *Filter) getFilterChanges() (interface{}, error) {
 	case pendingTxFilter:
 		// TODO
 	case logFilter:
-		// TODO
+		return f.getFilterLogs()
 	}
 
-	return nil, nil
+	return nil, errors.New("unsupported filter")
 }
 
-func (f *Filter) getFilterLogs() []*ethtypes.Log {
-	// TODO
-	return nil
-}
+func (f *Filter) getFilterLogs() ([]*ethtypes.Log, error) {
+	ret := []*ethtypes.Log{}
 
-func includes(addresses []common.Address, a common.Address) bool {
-	for _, addr := range addresses {
-		if addr == a {
-			return true
+	// filter specific block only
+	if f.blockHash != nil {
+		block, err := f.backend.GetBlockByHash(*f.blockHash, true)
+		if err != nil {
+			return nil, err
+		}
+
+		// if the logsBloom == 0, there are no logs in that block
+		if txs, ok := block["transactions"].([]common.Hash); !ok {
+			return ret, nil
+		} else if len(txs) != 0 {
+			return f.checkMatches(block)
 		}
 	}
 
-	return false
+	// filter range of blocks
+	num, err := f.backend.BlockNumber()
+	if err != nil {
+		return nil, err
+	}
+
+	// if f.fromBlock is set to 0, set it to the latest block number
+	if f.fromBlock == nil || f.fromBlock.Cmp(big.NewInt(0)) == 0 {
+		f.fromBlock = big.NewInt(int64(num))
+	}
+
+	// if f.toBlock is set to 0, set it to the latest block number
+	if f.toBlock == nil || f.toBlock.Cmp(big.NewInt(0)) == 0 {
+		f.toBlock = big.NewInt(int64(num))
+	}
+
+	log.Debug("[ethAPI] Retrieving filter logs", "fromBlock", f.fromBlock, "toBlock", f.toBlock,
+		"topics", f.topics, "addresses", f.addresses)
+
+	from := f.fromBlock.Int64()
+	to := f.toBlock.Int64()
+
+	for i := from; i <= to; i++ {
+		block, err := f.backend.GetBlockByNumber(NewBlockNumber(big.NewInt(i)), true)
+		if err != nil {
+			f.err = err
+			log.Debug("[ethAPI] Cannot get block", "block", block["number"], "error", err)
+			break
+		}
+
+		log.Debug("[ethAPI] filtering", "block", block)
+
+		// TODO: block logsBloom is often set in the wrong block
+		// if the logsBloom == 0, there are no logs in that block
+
+		if txs, ok := block["transactions"].([]common.Hash); !ok {
+			continue
+		} else if len(txs) != 0 {
+			logs, err := f.checkMatches(block)
+			if err != nil {
+				f.err = err
+				break
+			}
+
+			ret = append(ret, logs...)
+		}
+	}
+
+	return ret, nil
+}
+
+func (f *Filter) checkMatches(block map[string]interface{}) ([]*ethtypes.Log, error) {
+	transactions, ok := block["transactions"].([]common.Hash)
+	if !ok {
+		return nil, errors.New("invalid block transactions")
+	}
+
+	unfiltered := []*ethtypes.Log{}
+
+	for _, tx := range transactions {
+		logs, err := f.backend.GetTxLogs(common.BytesToHash(tx[:]))
+		if err != nil {
+			return nil, err
+		}
+
+		unfiltered = append(unfiltered, logs...)
+	}
+
+	return filterLogs(unfiltered, f.fromBlock, f.toBlock, f.addresses, f.topics), nil
 }
 
 // filterLogs creates a slice of logs matching the given criteria.
+// [] -> anything
+// [A] -> A in first position of log topics, anything after
+// [null, B] -> anything in first position, B in second position
+// [A, B] -> A in first position and B in second position
+// [[A, B], [A, B]] -> A or B in first position, A or B in second position
 func filterLogs(logs []*ethtypes.Log, fromBlock, toBlock *big.Int, addresses []common.Address, topics [][]common.Hash) []*ethtypes.Log {
 	var ret []*ethtypes.Log
 Logs:
@@ -193,4 +275,14 @@ Logs:
 		ret = append(ret, log)
 	}
 	return ret
+}
+
+func includes(addresses []common.Address, a common.Address) bool {
+	for _, addr := range addresses {
+		if addr == a {
+			return true
+		}
+	}
+
+	return false
 }